modified:   helpers/delete_files.js 	modified:   routes/category_router.js 	deleted:    uploads/image-1774789731738.png 	deleted:    uploads/image-1774877363009.png 	deleted:    uploads/image-1774877372181.png

+++ helpers/delete_files.js
@@ -20,6 +20,7 @@ function deleteFiles(filePaths) {
 }
 
 function deleteFile(filePath) {
+  if(!filePath) return
   fs.unlink(join('uploads',filePath), (err) => {
     if (err) {
       console.error("Error deleting file:", err);
